Added custom layouts, recycler view, adapter, details and api calls

+++ app/src/main/java/com/ritvikkar/weatherapp/data/Main.java
@@ -10,10 +10,10 @@ public class Main {
     public Double temp;
     @SerializedName("pressure")
     @Expose
-    public Integer pressure;
+    public Double pressure;
     @SerializedName("humidity")
     @Expose
-    public Integer humidity;
+    public Double humidity;
     @SerializedName("temp_min")
     @Expose
     public Double tempMin;
@@ -25,11 +25,11 @@ public class Main {
         return temp;
     }
 
-    public Integer getPressure() {
+    public Double getPressure() {
         return pressure;
     }
 
-    public Integer getHumidity() {
+    public Double getHumidity() {
         return humidity;
     }
